+++ src/DuckFunkCoreServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Torralbodavid\DuckFunkCore;
 
 use Illuminate\Support\ServiceProvider;
+use Laravel\Ui\UiServiceProvider;
 use Torralbodavid\DuckFunkCore\Models\Arcturus\User;
 use Torralbodavid\DuckFunkCore\Models\Housekeeping\News;
 use Torralbodavid\DuckFunkCore\Observers\NewsObserver;
@@ -16,7 +17,7 @@ class DuckFunkCoreServiceProvider extends ServiceProvider
     public function boot()
     {
         // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'duck-funk-core');
-        $this->loadViewsFrom(__DIR__.'/resources/views/'.config('duck-funk.template'), 'duck-funk-core');
+        $this->loadViewsFrom(__DIR__.'/resources/views', 'duck-funk-core');
         $this->loadViewsFrom(__DIR__.'/resources/views/housekeeping', 'housekeeping');
         $this->loadViewsFrom(__DIR__.'/resources/views/auth', 'auth');
 
@@ -28,12 +29,10 @@ class DuckFunkCoreServiceProvider extends ServiceProvider
             __DIR__.'/../config/config.php' => config_path('duck-funk.php'),
         ], 'duck-funk-core/config');
 
-        // Publishing the views.
-        /*$this->publishes([
-            __DIR__.'/resources/views' => resource_path('views/vendor/duck-funk-core'),
-        ], 'views');*/
+        $this->publishes([
+            __DIR__.'/resources/views' => resource_path('views/vendor/duck-funk-core/'),
+        ], 'duck-funk-core/views');
 
-        // Publishing assets.
         $this->publishes([
             __DIR__.'/../public' => public_path('vendor/duck-funk-core'),
         ], 'duck-funk-core/assets');
@@ -54,6 +53,7 @@ class DuckFunkCoreServiceProvider extends ServiceProvider
      */
     public function register()
     {
+        $this->app->register(UiServiceProvider::class);
         $this->app->register(DuckFunkEventServiceProvider::class);
         // Automatically apply the package configuration
         $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'duck-funk');