+++ src/resources/views/fuse/layouts/header.blade.php
@@ -371,18 +371,9 @@
                         <div class="col-sm-6">
                             <div class="float-right d-none d-md-block">
                                 <div class="dropdown">
-                                    <button class="btn btn-danger dropdown-toggle arrow-none waves-effect waves-light"
-                                            type="button" data-toggle="dropdown" aria-haspopup="true"
-                                            aria-expanded="false">
-                                        <i class="mdi mdi-settings mr-2"></i> Settings
-                                    </button>
-                                    <div class="dropdown-menu dropdown-menu-right">
-                                        <a class="dropdown-item" href="#">Action</a>
-                                        <a class="dropdown-item" href="#">Another action</a>
-                                        <a class="dropdown-item" href="#">Something else here</a>
-                                        <div class="dropdown-divider"></div>
-                                        <a class="dropdown-item" href="#">Separated link</a>
-                                    </div>
+                                    <a href="{{ core()->page('hotel') }}" class="btn btn-danger arrow-none waves-effect waves-light">
+                                        <i class="retro retro-mid-hotel-button-medium-icon"></i> Entra en el hotel
+                                    </a>
                                 </div>
                             </div>
                         </div> <!-- end col-->