+++ tests/TestCase.php
@@ -3,15 +3,27 @@
 namespace Torralbodavid\DuckFunkCore\Tests;
 
 use Torralbodavid\DuckFunkCore\DuckFunkCoreServiceProvider;
+use Torralbodavid\DuckFunkCore\Models\Arcturus\Permissions;
+use Torralbodavid\DuckFunkCore\Models\Arcturus\User;
 
 abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
+    protected User $user;
+
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->loadMigrationsFrom(__DIR__.'/../tests/database/migrations');
         $this->withFactories(__DIR__.'/../database/factories');
+
+        $this->user = factory(User::class)->create();
+        $this->permissions = factory(Permissions::class)->create(['id' => 1, 'rank_name' => 'Test']);
+    }
+
+    protected function makeAuth()
+    {
+        return $this->be($this->user);
     }
 
     protected function getPackageProviders($app)
@@ -21,6 +33,7 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
     protected function getEnvironmentSetUp($app)
     {
+        $app['config']->set('auth.providers.users.model', User::class);
         $app['config']->set('database.default', 'testbench');
         $app['config']->set('database.connections.testbench', [
             'driver'   => 'sqlite',