+++ tests/database/migrations/2020_04_30_000000_add_bans_test_table.php
@@ -1,0 +1,28 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddBansTestTable extends \Illuminate\Database\Migrations\Migration
+{
+    public function up()
+    {
+        Schema::create('bans', function (Blueprint $table) {
+            $table->unsignedInteger('id', true);
+            $table->unsignedInteger('user_id');
+            $table->string('ip');
+            $table->string('machine_id');
+            $table->integer('user_staff_id');
+            $table->integer('timestamp');
+            $table->integer('ban_expire');
+            $table->string('ban_reason');
+            $table->enum('type', ['account', 'ip', 'machine', 'super']);
+            $table->integer('cfh_topic');
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('bans');
+    }
+}