+++ tests/database/migrations/2020_05_01_000000_add_messenger_friendrequests_test_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddMessengerFriendrequestsTestTable extends \Illuminate\Database\Migrations\Migration
+{
+    public function up()
+    {
+        Schema::create('messenger_friendrequests', function (Blueprint $table) {
+            $table->unsignedInteger('id', true);
+            $table->unsignedInteger('user_to_id');
+            $table->unsignedInteger('user_from_id');
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('messenger_friendrequests');
+    }
+}