+++ tests/database/migrations/2020_05_01_000000_add_messenger_friendships_test_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddMessengerFriendshipsTestTable extends \Illuminate\Database\Migrations\Migration
+{
+    public function up()
+    {
+        Schema::create('messenger_friendships', function (Blueprint $table) {
+            $table->unsignedInteger('id', true);
+            $table->unsignedInteger('user_one_id');
+            $table->unsignedInteger('user_two_id');
+            $table->unsignedInteger('relation');
+            $table->unsignedInteger('friends_since');
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('messenger_friendrequests');
+    }
+}